requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
sentence-transformers
faiss-cpu
requests
numpy
